test: 테스트 분리 및 중복 코드 삭제
